updating synthetic validation test exception to ValidationError. refactor to synthetic_agentic core to reside inside new datagen/synthetic_agentic dir
