+++ tests/required/datagen/test_tfut_synthetic_graph.py
@@ -17,8 +17,8 @@
 from typing import cast
 
 from inference_perf.datagen.replay.replay_graph_types import ReplayGraph
-from inference_perf.datagen.synthetic_agentic import build_graph_for_session
-from inference_perf.datagen.synthetic_themes import GENERIC_THEME
+from inference_perf.datagen.synthetic_agentic.synthetic_agentic_datagen import build_graph_for_session
+from inference_perf.datagen.synthetic_agentic.synthetic_themes import GENERIC_THEME
 from inference_perf.datagen.replay.otel_trace_to_replay_graph import tag_user_facing_events
 from inference_perf.config.common import Distribution
 from inference_perf.config.datagen.replay import SyntheticAgenticConfig
